+++ bin/cli.js
@@ -50,6 +50,9 @@ ghauth(ghauthOpts, function (err, auth) {
     // merge defaults and command line arguments into options
 
     extend(options, defaults, argv)
+
+    // filter options through whitelist
+
     var whitelist = ghRelease.OPTIONS.whitelist
 
     Object.keys(options).forEach(function (key) {