+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 This change log adheres to [keepachangelog.com](http://keepachangelog.com).
 
+## Unreleased
+### Changed
+* moved all logging and CLI-related logic out of API and into CLI
+* **breaking:** changed API parameters from `options, auth, callback` to `options, callback`
+  * `auth` is now in `options` as `options.auth`
+
 ## [1.1.8] - 2015-04-22
 ### Fixed
 * exit process with code 1 on aborted release or invalid directory ([#31](https://github.com/ngoldman/gh-release/issues/31))